+++ inst/doc/Using_predictInterval.R
@@ -3,7 +3,7 @@ library(ggplot2); library(knitr); library(merTools)
 knitr::opts_chunk$set(
   cache=FALSE,
   comment="#>",
-  collapse=TRUE, 
+  collapse=TRUE,
   echo = TRUE
 )
 
@@ -15,7 +15,7 @@ display(fm1)
 
 ## ----predInt-------------------------------------------------------------
 PI.time <- system.time(
-  PI <- predictInterval(merMod = fm1, newdata = sleepstudy, 
+  PI <- predictInterval(merMod = fm1, newdata = sleepstudy,
                         level = 0.95, n.sims = 1000,
                         stat = "median", type="linear.prediction",
                         include.resid.var = TRUE)
@@ -27,9 +27,32 @@ kable(head(PI))
 ## ----Inspect predInt 2, fig.width=7, fig.align="center"------------------
 library(ggplot2);
 ggplot(aes(x=1:30, y=fit, ymin=lwr, ymax=upr), data=PI[1:30,]) +
-  geom_point() + 
+  geom_point() +
   geom_linerange() +
   labs(x="Index", y="Prediction w/ 95% PI") + theme_bw()
+
+## ----prepSleep, message=FALSE, warning=FALSE-----------------------------
+fm1 <- lmer(Reaction ~ Days + (Days|Subject), data=sleepstudy)
+display(fm1)
+
+## ----showSleep, message=FALSE, warning=FALSE-----------------------------
+sleepstudy[1,]
+predictInterval(fm1, sleepstudy[1,], include.resid.var=0) #predict the average body fat for a group of 196cm female baseball players
+
+## ----ShowFat2, message=FALSE, warning=FALSE------------------------------
+predictInterval(fm1, sleepstudy[1,], include.resid.var=0, ignore.fixed.terms = 1) 
+# predict the average reaction time for a subject at day 0, taking the global intercept 
+# (mean reaction time) as fully known
+predictInterval(fm1, sleepstudy[1,], include.resid.var=0, ignore.fixed.terms = "(Intercept)") 
+#Same as above
+predictInterval(fm1, sleepstudy[1,], include.resid.var=0, ignore.fixed.terms = 1:2) 
+# as above, taking the first two fixed effects (intercept and days effect) as fully known
+
+## ----ShowFat3, message=FALSE, warning=FALSE------------------------------
+predictInterval(fm1, sleepstudy[1,], include.resid.var=0, 
+                fix.intercept.variance = TRUE) 
+# predict the average reaction time for a subject at day 0,, using an ad-hoc 
+# correction for the covariance of the intercept with the random intercept effects.
 
 ## ----arm.Sim, fig.width=7, fig.height=4, fig.align="center"--------------
 PI.arm.time <- system.time(
@@ -46,13 +69,13 @@ comp.data <- rbind(data.frame(Predict.Method="predictInterval()", x=(1:nrow(PI))
                    data.frame(Predict.Method="arm::sim()", x=(1:nrow(PI.arm))+0.1, PI.arm))
 
 ggplot(aes(x=x, y=fit, ymin=lwr, ymax=upr, color=Predict.Method), data=comp.data[c(1:30,181:210),]) +
-  geom_point() + 
+  geom_point() +
   geom_linerange() +
   labs(x="Index", y="Prediction w/ 95% PI") +
   theme_bw() +  theme(legend.position="bottom") +
   scale_color_brewer(type = "qual", palette = 2)
-          
-              
+
+
 
 ## ----bootMer.1, fig.width=7, fig.height=4, fig.align="center"------------
 ##Functions for bootMer() and objects
@@ -72,7 +95,7 @@ sumBoot <- function(merBoot) {
 
 ##lme4::bootMer() method 1
 PI.boot1.time <- system.time(
-  boot1 <- lme4::bootMer(fm1, mySumm, nsim=1000, use.u=FALSE, type="parametric")
+  boot1 <- lme4::bootMer(fm1, mySumm, nsim=250, use.u=FALSE, type="parametric")
 )
 
 PI.boot1 <- sumBoot(boot1)
@@ -81,17 +104,17 @@ comp.data <- rbind(data.frame(Predict.Method="predictInterval()", x=(1:nrow(PI))
                    data.frame(Predict.Method="lme4::bootMer() - Method 1", x=(1:nrow(PI.boot1))+0.1, PI.boot1))
 
 ggplot(aes(x=x, y=fit, ymin=lwr, ymax=upr, color=Predict.Method), data=comp.data[c(1:30,181:210),]) +
-  geom_point() + 
+  geom_point() +
   geom_linerange() +
   labs(x="Index", y="Prediction w/ 95% PI") +
   theme_bw() +  theme(legend.position="bottom") +
   scale_color_brewer(type = "qual", palette = 2)
-  
+
 
 ## ----bootMer.2, fig.width=7, fig.height=4, fig.align="center"------------
 ##lme4::bootMer() method 2
 PI.boot2.time <- system.time(
-  boot2 <- lme4::bootMer(fm1, mySumm, nsim=1000, use.u=TRUE, type="parametric")
+  boot2 <- lme4::bootMer(fm1, mySumm, nsim=250, use.u=TRUE, type="parametric")
 )
 
 PI.boot2 <- sumBoot(boot2)
@@ -100,18 +123,18 @@ comp.data <- rbind(data.frame(Predict.Method="predictInterval()", x=(1:nrow(PI))
                    data.frame(Predict.Method="lme4::bootMer() - Method 2", x=(1:nrow(PI.boot2))+0.1, PI.boot2))
 
 ggplot(aes(x=x, y=fit, ymin=lwr, ymax=upr, color=Predict.Method), data=comp.data[c(1:30,181:210),]) +
-  geom_point() + 
+  geom_point() +
   geom_linerange() +
   labs(x="Index", y="Prediction w/ 95% PI") +
   theme_bw() +  theme(legend.position="bottom") +
   scale_color_brewer(type = "qual", palette = 2)
-  
-  
+
+
 
 ## ----bootMer.3, fig.width=7, fig.height=4, fig.align="center"------------
 ##lme4::bootMer() method 3
 PI.boot3.time <- system.time(
-  boot3 <- lme4::bootMer(fm1, mySumm, nsim=1000, use.u=TRUE, type="semiparametric")
+  boot3 <- lme4::bootMer(fm1, mySumm, nsim=250, use.u=TRUE, type="semiparametric")
 )
 
 PI.boot3 <- sumBoot(boot3)
@@ -120,13 +143,13 @@ comp.data <- rbind(data.frame(Predict.Method="predictInterval()", x=(1:nrow(PI))
                    data.frame(Predict.Method="lme4::bootMer() - Method 3", x=(1:nrow(PI.boot3))+0.1, PI.boot3))
 
 ggplot(aes(x=x, y=fit, ymin=lwr, ymax=upr, color=Predict.Method), data=comp.data[c(1:30,181:210),]) +
-  geom_point() + 
+  geom_point() +
   geom_linerange() +
   labs(x="Index", y="Prediction w/ 95% PI") +
   theme_bw() +  theme(legend.position="bottom") +
   scale_color_brewer(type = "qual", palette = 2)
-  
-  
+
+
 
 ## ----echo=FALSE, message=FALSE-------------------------------------------
 library(rstanarm)
@@ -145,9 +168,9 @@ central_intervals <- function(x, prob) {
 
 ## ---- message=FALSE, fig.width=7, fig.height=4, fig.align="center"-------
 PI.time.stan <- system.time({
-  fm_stan <- stan_lmer(Reaction ~ Days + (Days|Subject), data = sleepstudy, 
-                       verbose = FALSE, open_progress = FALSE, refresh = -1, 
-                       show_messages=FALSE)
+  fm_stan <- stan_lmer(Reaction ~ Days + (Days|Subject), data = sleepstudy,
+                       verbose = FALSE, open_progress = FALSE, refresh = -1,
+                       show_messages=FALSE, chains = 1)
   zed <- posterior_predict(fm_stan)
   PI.stan <- cbind(apply(zed, 2, median), central_intervals(zed, prob=0.95))
 })
@@ -162,7 +185,7 @@ comp.data <- rbind(data.frame(Predict.Method="predictInterval()", x=(1:nrow(PI))
                    data.frame(Predict.Method="rstanArm", x=(1:nrow(PI.stan))+0.1, PI.stan))
 
 ggplot(aes(x=x, y=fit, ymin=lwr, ymax=upr, color=Predict.Method), data=comp.data[c(1:30,181:210),]) +
-  geom_point() + 
+  geom_point() +
   geom_linerange() +
   labs(x="Index", y="Prediction w/ 95% PI") +
   theme_bw() +  theme(legend.position="bottom") +
@@ -170,7 +193,7 @@ ggplot(aes(x=x, y=fit, ymin=lwr, ymax=upr, color=Predict.Method), data=comp.data
 
 
 ## ---- echo=FALSE---------------------------------------------------------
-times <- rbind(PI.time, PI.arm.time, PI.boot1.time, PI.boot2.time, PI.boot3.time, 
+times <- rbind(PI.time, PI.arm.time, PI.boot1.time, PI.boot2.time, PI.boot3.time,
                PI.time.stan)[, 1:3]
 rownames(times) <- c("predictInterval()", "arm::sim()", "lme4::bootMer()-Method 1", "lme4::bootMer()-Method 2", "lme4::bootMer()-Method 3", "rstanarm:predict")
 kable(times)